extract request-context core microcrate
