+++ crates/bitnet-request-context-core/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "bitnet-request-context-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Reusable request context metadata contracts for inference pipelines"
+rust-version.workspace = true
+
+[dependencies]
+
+[dev-dependencies]
+
+[lints]
+workspace = true
